// frontend/src/shared/types/entities.ts
export type ID = string;

export interface Pagination {
    page: number;
    limit: number;
    total: number;
    pages: number;
}

export interface PaginatedResponse<T> {
    data: T[];
    pagination: Pagination;
}

export interface Customer {
    _id: ID;
    user: ID;
    name: string;
    email?: string;
    phone?: string;
    company?: string;
    address?: string;
    createdAt: string;
    updatedAt: string;
    __v?:number; 
}

export type CustomerCreate = Pick<Customer, "name" | "email" | "phone" | "company" | "address">;
export type CustomerUpdate = Partial<Omit<CustomerCreate, never>>;

export interface LineItem {
    description: string;
    quantity: number;
    unitPrice: number;
    taxRate?: number; // (0-100%)
}

export type QuoteStatusUserSettable = "draft" | "sent" | "accepted" | "declined";
export type QuoteStatus = QuoteStatusUserSettable | "expired" | "converted";

export interface Quote {
    _id: ID;
    user: ID;
    customer: ID | Customer;
    quoteNumber: string;
    issueDate: string;
    expiryDate?: string;
    items: LineItem[];
    total: number;
    status: QuoteStatus;
    notes?: string;
    createdAt?: string;
    updatedAt?: string;
}

export interface QuoteCreate {
    customer: ID;
    quoteNumber: string;
    issueDate: string;
    expiryDate?: string;
    status: QuoteStatus;
    items: LineItem[];
    notes?: string;
}

export type QuoteUpdate = Partial<Pick<Quote, "customer" | "quoteNumber" | "status" | "items" | "notes" | "issueDate" | "expiryDate">>;

export type InvoiceStatus = "draft" | "sent" | "partially_paid" | "paid" | "overdue";

export interface Invoice {
    _id: ID;
    user?: ID;
    customer: ID | Customer;
    invoiceNumber: string;
    issueDate: string;
    dueDate: string;
    items: LineItem[];
    status: InvoiceStatus;
    notes?: string;
    createdAt?: string;
    updatedAt?: string;
    quote?: string | Quote;
    totals: {
        subtotal: number;
        tax: number;
        total: number;
    };
}

export interface InvoiceCreate {
    customer: ID;
    invoiceNumber: string;
    issueDate: string;
    dueDate: string;
    items: LineItem[];
    notes?: string;
    quote?: string;
}

export type InvoiceUpdate = Partial<Pick<Invoice, "items" | "notes" | "dueDate">>;

export type PaymentMethod = "bank_transfer" | "card" | "cash" | "paypal";
export type PaymentStatus = "completed" | "failed" | "pending";

export interface Payment {
    _id: ID;
    paymentId?: string;
    user?: ID;
    invoice: ID | (Invoice & { customer? : Customer });
    amount: number;
    paymentMethod: PaymentMethod;
    paymentDate?: string;
    dueDate?: string;
    status: PaymentStatus;
    notes?: string;
    createdAt?: string;
    updatedAt?: string;
}

export interface PaymentCreate {
    paymentId: string;
    invoice: ID;
    amount: number;
    paymentMethod: PaymentMethod;
    paymentDate?: string;
    dueDate?: string;
    notes?: string;
}

export type PaymentUpdate = Partial<Pick<Payment, "amount" | "paymentMethod" | "paymentDate" | "dueDate" | "notes">>;

/* ------------- Components  ------------- */

// Dashboard
export interface SummaryCardProps {
    title: string;
    subtitle: string;
    value: number | string;
    loading: boolean;
    color: string;
}

export interface StatPrevCardItem {
    status: string;
    percentage: number;
}

export interface StatPrevCardProps {
    title: string;
    preview?: StatPrevCardItem[];
    loading: boolean;
}